Move functions to utils and consts to dictionaries
