+++ apps/season-editor/src/pages/components/styled-multiline-werapper.tsx
@@ -1,0 +1,7 @@
+import { styled } from "@mui/material";
+
+export const StyledMultilineInputWrapper = styled("div")(({ theme }) => ({
+  display: "grid",
+  gridTemplateColumns: "1fr 1fr 1fr",
+  gap: theme.spacing(2),
+}));
