+++ apps/season-editor/src/pages/dictionaries/available-languages.ts
@@ -1,0 +1,3 @@
+import { LanguageCode } from "../types";
+
+export const availableLanguages: LanguageCode[] = ["de", "en"];
