+++ apps/season-editor/src/pages/dictionaries/default-language.ts
@@ -1,0 +1,3 @@
+import { LanguageCode } from "../types";
+
+export const DEFAULT_LANGUAGE: LanguageCode = "de";
