+++ apps/season-editor/src/pages/dictionaries/index.ts
@@ -1,0 +1,2 @@
+export * from "./available-languages";
+export * from "./default-language";
