+++ apps/season-editor/src/pages/season-editor.tsx
@@ -12,7 +12,7 @@ import {
 } from "@mui/material";
 import { FormProvider, useForm } from "react-hook-form";
 import { parse, stringify } from "yaml";
-import { LocalizedString, SeasonDefinition } from "@talentdigital/season";
+import { SeasonDefinition } from "@talentdigital/season";
 import { ChangeEvent, useEffect, useState } from "react";
 
 import {
@@ -23,12 +23,14 @@ import {
   SectionVisibilityButton,
   TestItems,
 } from "./components";
-import { FormInputs, SectionName } from "./types";
-
-type LanguageCode = `${keyof LocalizedString}`;
-
-const DEFAULT_LANGUAGE: LanguageCode = "en"; // TODO: change to "de"
-const availableLanguages: LanguageCode[] = ["de", "en"];
+import { FormInputs, LanguageCode, SectionName } from "./types";
+import {
+  detectLanguage,
+  extractEpisodes,
+  extractFromCompetences,
+  mapToSeasonObject,
+} from "./utils";
+import { DEFAULT_LANGUAGE, availableLanguages } from "./dictionaries";
 
 /* TODO:
   - Allow to init without importing a file
@@ -93,11 +95,8 @@ export const SeasonEditor = () => {
     if (!season) return;
 
     const episodes = extractEpisodes(season, language);
-
-    const { competenceAreas, competences, subCompetences } = extractCompetences(
-      season,
-      language
-    );
+    const { competenceAreas, competences, subCompetences, testItems } =
+      extractFromCompetences(season, language);
 
     reset({
       title: season.title[language] ?? "",
@@ -108,6 +107,7 @@ export const SeasonEditor = () => {
       ...competences,
       ...subCompetences,
       episodes,
+      testItems,
     });
   }, [season, language, reset]);
 
@@ -248,239 +248,6 @@ export const SeasonEditor = () => {
     </Box>
   );
 };
-
-function detectLanguage(season?: SeasonDefinition): LanguageCode {
-  if (typeof season?.title === "object") {
-    const maybeLanguageKey = Object.keys(season.title)[0];
-
-    if (isLanguageKey(maybeLanguageKey)) {
-      return maybeLanguageKey;
-    }
-  }
-
-  return DEFAULT_LANGUAGE;
-}
-
-function isLanguageKey(key: string): key is LanguageCode {
-  return availableLanguages.includes(key as LanguageCode);
-}
-
-function mapToSeasonObject(
-  originalFileLoaded: SeasonDefinition,
-  valuesWithEmpty: FormInputs,
-  language: LanguageCode
-) {
-  const newFile: SeasonDefinition = JSON.parse(
-    JSON.stringify(originalFileLoaded)
-  );
-  const values: Partial<FormInputs> = Object.fromEntries(
-    Object.entries(valuesWithEmpty).filter((entries) => Boolean(entries[1]))
-  ) as Partial<FormInputs>;
-
-  newFile.title[language] = values.title;
-  newFile.info[language] = values.info;
-  newFile.assetsURL = values.assetsURL;
-  newFile.seasonEndMessage[language] = values.seasonEndMessage;
-  newFile.competenceAreas = mapFormCompetenceAreasToSeasonCompetenceAreas(
-    values,
-    newFile,
-    language
-  );
-  newFile.episodes = mapFormEpisodesToSeasonEpisodes(values, newFile, language);
-
-  return newFile;
-}
-
-function mapFormCompetenceAreasToSeasonCompetenceAreas(
-  values: Partial<FormInputs>,
-  oldValues: SeasonDefinition,
-  language: LanguageCode
-): SeasonDefinition["competenceAreas"] {
-  return (
-    values.competenceAreas?.reduce((accumulator, competenceArea) => {
-      return {
-        ...accumulator,
-        [competenceArea.competenceAreaId]: {
-          name: {
-            ...oldValues.competenceAreas[competenceArea.competenceAreaId]?.name,
-            [language]: competenceArea.name,
-          },
-          competences: mapFormCompetencesToSeasonCompetences(
-            values,
-            oldValues,
-            language,
-            competenceArea.competenceAreaId
-          ),
-        },
-      };
-    }, {}) ?? {}
-  );
-}
-
-function mapFormCompetencesToSeasonCompetences(
-  values: Partial<FormInputs>,
-  oldValues: SeasonDefinition,
-  language: LanguageCode,
-  competenceAreaId: string
-) {
-  const competences = values[`competences-${competenceAreaId}`];
-
-  return (
-    competences?.reduce((accumulator, competence) => {
-      return {
-        ...accumulator,
-        [competence.competenceId]: {
-          name: {
-            ...oldValues.competenceAreas[competence.competenceId]?.name,
-            [language]: competence.name,
-          },
-          subCompetences: mapFormSubCompetencesToSeasonSubCompetences(
-            values,
-            oldValues,
-            language,
-            competenceAreaId,
-            competence.competenceId
-          ),
-        },
-      };
-    }, {}) ?? {}
-  );
-}
-
-function mapFormSubCompetencesToSeasonSubCompetences(
-  values: Partial<FormInputs>,
-  oldValues: SeasonDefinition,
-  language: LanguageCode,
-  competenceAreaId: string,
-  competenceId: string
-) {
-  const subCompetences =
-    values[`subCompetences-${competenceAreaId}-${competenceId}`];
-
-  return (
-    subCompetences?.reduce((accumulator, subCompetence) => {
-      return {
-        ...accumulator,
-        [subCompetence.subCompetenceId]: {
-          name: {
-            ...oldValues.competenceAreas[competenceAreaId]?.competences[
-              competenceId
-            ]?.subCompetences[subCompetence.subCompetenceId]?.name,
-            [language]: subCompetence.name,
-          },
-        },
-      };
-    }, {}) ?? {}
-  );
-}
-
-function mapFormEpisodesToSeasonEpisodes(
-  values: Partial<FormInputs>,
-  oldValues: SeasonDefinition,
-  language: LanguageCode
-): SeasonDefinition["episodes"] {
-  return (
-    values.episodes?.reduce((accumulator, episode) => {
-      return {
-        ...accumulator,
-        [episode.episodeId]: {
-          title: {
-            ...oldValues.episodes[episode.episodeId]?.title,
-            [language]: episode.title,
-          },
-          description: {
-            ...oldValues.episodes[episode.episodeId]?.description,
-            [language]: episode.description,
-          },
-          maturity: episode.maturity,
-          imageUrl: episode.imageUrl,
-          format: episode.format,
-          formatConfiguration: episode.formatConfiguration,
-        },
-      };
-    }, {}) ?? {}
-  );
-}
-
-function extractEpisodes(
-  season: SeasonDefinition,
-  language: LanguageCode
-): FormInputs["episodes"] {
-  return Object.entries(season.episodes).map(([episodeId, episodeValue]) => ({
-    episodeId,
-    title: episodeValue.title[language] ?? "",
-    description: episodeValue.description[language] ?? "",
-    maturity: episodeValue.maturity,
-    imageUrl: episodeValue.imageUrl ?? "",
-    format: episodeValue.format,
-    formatConfiguration: episodeValue.formatConfiguration ?? "",
-  }));
-}
-
-type ExtractCompetencesReturn = {
-  competenceAreas: FormInputs["competenceAreas"];
-  competences: Record<
-    `competences-${string}`,
-    FormInputs[`competences-${string}`]
-  >;
-  subCompetences: Record<
-    `subCompetences-${string}-${string}`,
-    FormInputs[`subCompetences-${string}-${string}`]
-  >;
-};
-
-function extractCompetences(
-  season: SeasonDefinition,
-  language: LanguageCode
-): ExtractCompetencesReturn {
-  const competenceAreas: ExtractCompetencesReturn["competenceAreas"] = [];
-  const competences: ExtractCompetencesReturn["competences"] = {};
-  const subCompetences: ExtractCompetencesReturn["subCompetences"] = {};
-
-  Object.entries(season.competenceAreas).forEach(
-    ([competenceAreaId, competenceAreaValue]) => {
-      competenceAreas.push({
-        competenceAreaId,
-        name: competenceAreaValue.name?.[language] ?? "",
-      });
-
-      Object.entries(competenceAreaValue.competences).forEach(
-        ([competenceId, competenceValue]) => {
-          const competenceKey = `competences-${competenceAreaId}` as const;
-          if (!competences[competenceKey]) {
-            competences[competenceKey] = [];
-          }
-
-          competences[competenceKey].push({
-            competenceAreaId,
-            competenceId,
-            name: competenceValue.name?.[language] ?? "",
-          });
-
-          Object.entries(competenceValue.subCompetences).forEach(
-            ([subCompetenceId, subCompetenceValue]) => {
-              const subCompetenceKey =
-                `subCompetences-${competenceAreaId}-${competenceId}` as const;
-
-              if (!subCompetences[subCompetenceKey]) {
-                subCompetences[subCompetenceKey] = [];
-              }
-
-              subCompetences[subCompetenceKey].push({
-                competenceAreaId,
-                competenceId,
-                subCompetenceId,
-                name: subCompetenceValue.name?.[language] ?? "",
-              });
-            }
-          );
-        }
-      );
-    }
-  );
-
-  return { competenceAreas, competences, subCompetences };
-}
 
 const StyledNavigation = styled("div")(({ theme }) => ({
   padding: theme.spacing(2),
