+++ apps/season-editor/src/pages/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./form-inputs";
+export * from "./language-code";
 export * from "./section-name";
