+++ apps/season-editor/src/pages/types/language-code.ts
@@ -1,0 +1,3 @@
+import { LocalizedString } from "@talentdigital/season";
+
+export type LanguageCode = `${keyof LocalizedString}`;
